refactor: moved enums to package
