+++ src/main/java/cn/kastner/oj/domain/enums/ContestStatus.java
@@ -1,4 +1,4 @@
-package cn.kastner.oj.domain;
+package cn.kastner.oj.domain.enums;
 
 public enum ContestStatus {
   NOT_STARTED,
